Move analyse to GTRepository+Merging
