+++ ObjectiveGit/GTRepository+Pull.h
@@ -7,22 +7,8 @@
 //
 
 #import "GTRepository.h"
-#import "git2/merge.h"
 
 NS_ASSUME_NONNULL_BEGIN
-
-/// UserInfo key for conflicted files when pulling fails with a merge conflict
-extern NSString * const GTPullMergeConflictedFiles;
-
-/// An enum describing the result of the merge analysis.
-/// See `git_merge_analysis_t`.
-typedef NS_OPTIONS(NSInteger, GTMergeAnalysis) {
-	GTMergeAnalysisNone = GIT_MERGE_ANALYSIS_NONE,
-	GTMergeAnalysisNormal = GIT_MERGE_ANALYSIS_NORMAL,
-	GTMergeAnalysisUpToDate = GIT_MERGE_ANALYSIS_UP_TO_DATE,
-	GTMergeAnalysisUnborn = GIT_MERGE_ANALYSIS_UNBORN,
-	GTMergeAnalysisFastForward = GIT_MERGE_ANALYSIS_FASTFORWARD,
-};
 
 typedef void (^GTRemoteFetchTransferProgressBlock)(const git_transfer_progress *progress, BOOL *stop);
 
@@ -43,16 +29,6 @@ typedef void (^GTRemoteFetchTransferProgressBlock)(const git_transfer_progress *
 /// Returns YES if the pull was successful, NO otherwise (and `error`, if provided,
 /// will point to an error describing what happened).
 - (BOOL)pullBranch:(GTBranch *)branch fromRemote:(GTRemote *)remote withOptions:(nullable NSDictionary *)options error:(NSError **)error progress:(nullable GTRemoteFetchTransferProgressBlock)progressBlock;
-
-/// Analyze which merge to perform.
-///
-/// analysis   - The resulting analysis.
-/// fromBranch - The branch to merge from.
-/// error      - The error if one occurred. Can be NULL.
-///
-/// Returns YES if the analysis was successful, NO otherwise (and `error`, if provided,
-/// will point to an error describing what happened).
-- (BOOL)analyzeMerge:(GTMergeAnalysis *)analysis fromBranch:(GTBranch *)fromBranch error:(NSError **)error;
 
 @end
